objects/ember_emitter: merge ember emitter control in TRX
