+++ src/tr2/meson.build
@@ -196,7 +196,6 @@ sources = [
   'game/objects/traps/dart.c',
   'game/objects/traps/dart_emitter.c',
   'game/objects/traps/dying_monk.c',
-  'game/objects/traps/ember_emitter.c',
   'game/objects/traps/falling_ceiling.c',
   'game/objects/traps/flame_emitter.c',
   'game/objects/traps/gondola.c',
